Fix incorrect highlighting of after-EOL matches under the cursor.
Previously, search matches past the end of the line (for example, ^ or $
on a line containing "\n") would not be highlighted on the block
underneath the cursor. In rightleft mode, the same search would cause
odd screen corruption.

The buggy behavior was introduced in vim/vim@a443af8. When a highlight
goes past the end of a line, win_line() decrements col, which is later
passed as an endcol to screen_line(). This causes the matching cell to
be cleared instead of highlighted.

This change removes the special casing of these matches under the
cursor.