feat: rename 'diffall' to 'diffnomaster'
This is to avoid confusion, as some users may expect :diffall to enable
diff on all windows like :windo diffthis. I could just program it to
actually do that if there aren't any diff windows visible (as then
that's most likely what the user thought it did) but that could cause
even more confusion, e.g. if a user later tries to use it when some win-
dows are already diff windows and it doesn't do what they expect.