GvimExt: Create a submenu for "Edit with existing Vim"
If we have a lot of gVim instances, many numbers of "Edit with existing Vim"
items will be shown in the context menu with the current implementation.

E.g.

    Edit with Vim
    Edit with existing Vim - instance 1
    Edit with existing Vim - instance 2
    Edit with existing Vim - instance 3
    Edit with existing Vim - instance 4
    Edit with existing Vim - instance 5
    Edit with existing Vim - instance 6
    Edit with existing Vim - instance 7

I feel this is a bit noisy.

This PR makes the menu items structured and move the instances into a
submenu.  So the context menu becomes simpler.

E.g.

    Edit with Vim
    Edit with existing Vim > instance 1
                             instance 2
                             instance 3
                             instance 4
                             instance 5
                             instance 6
                             instance 7

If there is only one gVim instance, the item will be shown directly:

    Edit with Vim
    Edit with existing Vim - instance 1