dosinst: Use $USERPROFILE as a fallback of plugin directories
Creating plugin directories in $VIM as a fallback of $HOME is not useful.

If the install.exe is executed from the NSIS installer, $VIM is normally
`C:\Program Files (x86)\Vim`.  Normal users cannot put plugin files into
it, because the directory requires administrator privileges.

Currently, the NSIS installer has the following two options:
* [x] Create plugin directories in HOME or VIM
* [ ] Create plugin directories in VIM

The first one is the default, however, most people might not set $HOME.
Then the plugin directories will be created in $VIM (which is not useful).
If the user really wants to create in $VIM, he should select the second
one.

If the user doesn't set $HOME, it should fallback to $USERPROFILE.