Avoid calling ui_breakcheck recursively
It's possible to get stuck in an infinitely recursive call loop when
stderr is redirected.  A call to ui_breakcheck() will eventually make
its way to fill_input_buf(), which calls settmode() when stdin isn't a
tty.

settmode() calls down to vgetorpeek(), which then calls ui_breakcheck()
to see if it should break out of an infinite for loop.