Add full 64/32 bit range to -o and output
xxd is unable to output addresses higher that 0xffffffff even when compiled for a 64 bit architecture. This makes it hard to work with any files bigger than 4.3 GB or virtual files (like process memory) that can be addressed higher. It's odd limit since this limit is not imposed on seek and if custom input is feed to xxd -r with addresses higher than 0xffffffff it works fine. So parsing it is no problem.

This commit is removes this limitation allowing for the argument to -o to span the entire [-0xffffffffffffffff, +0xffffffffffffffff] range in case of 64 bit architecture or [-0xffffffff, +0xffffffff] in case of 32 bit architecture, same with the output generated by the program. To be as non-destructive as possible the output stays unchanged until 0xffffffff is exceeded in which case any additional digits is added as needed pushing everything over by however many digits.

The seek is still limited to the size of a signed long, partly due to the API used. Nothing changed regards to that. Testing showed no issues with both 32 and 64 bit binaries on my Linux 64 bit machine.