Avoided NULL pointer arithmetic
Arithmetic with NULL pointers, like in the definition of char *cp, are undefined behaviour and thus should be avoided. In addition, I could not find, wether strnicmp can deal with NULL arguments and delayed its execution to prevent the rather hackish approach "second or condition is only evaluated if first one is false".