Fix macOS fcntl/fsync usage properly
When fcntl fails, don't check `errno` when deciding whether to call
fsync, as there are different possible `errno` results and they are not
documented in the man page. Instead, always fall back to fsync. Worst
case is that it's only slightly slower but that should only happen when
something is wrong anyway so that's not a big deal.