Fix incorrect reference count in python layer
WHen calling PyObject_CallFunction with the "N" argument, the reference
count of the passed objects are not incremented. The subsequent
Py_DECREF(paths) causes an abort when compiling against debug python.

The "N" style format argument is for passing objects returned directory
from factory methods, so is not appropriate here.