DirectWrite: Force redraw when render target is recreated
DirectWrite sometimes fails with the `D2DERR_RECREATE_TARGET` error.  In
that case, we need to recreate the Direct2D render target.  We already
do it, however, just recreating it is not enough, and the Vim screen
becomes blank.  Typing CTRL-L can work around the problem, but it is not
convenient.  This patch automatically redraws the screen when the render
target is recreated.