Fix broken listchars handling for nbsp and space
Previously, `listchars` didn't properly handle characters like ' ゙' which
has a composing character 0x20 which is the same as space. As a result
the character would have turned into the listchar for space when
`listchars` was set.

PR #4046 attempted to fix this issue but it split the branches but
didn't address the fact that the composing char wasn't correctly
handled, and also made nbsp listchars no longer get syntax highlighting,
and unicode characters no longer worked for space listchars, e.g. →.

Fix the issue by reverting back to the original listchars handling code,
but check for whether the space character (0x20) that we are handling is
a composing char or an actual space.