win32: Use wide entry points
The patch 8.1.1091 enables to support Unicode environment variables
inside Vim. However, we still cannot access external Unicode env var.
E.g.:

```
> set a=Ā
> gvim --clean --cmd "set enc=utf-8"
:echo $a
?  " should be Ā
```

This is because we use ANSI entry points: WinMain() for GUI and main()
for CUI. To fix this, we need to use wide entry points: wWinMain() for
GUI and wmain() for CUI.

This also moves the entry point for CUI from main.c to os_w32exe.c to
make the code simple.

Note: This requires the `-municode` option to MinGW, which means now
MinGW-w64 is needed.