Don't spellcheck symbols such as emojis
Previously various "emoji" symbols would get marked as bad by the spell
checker as it didn't recognize it as a non-word character.

Examples: 🚀, 🇳🇿(should render as New Zealand flag), ♔, and a whole
bunch more.

This adds the following blocks from the supplementary multilingual
plane:

	https://codepoints.net/byzantine_musical_symbols
	https://codepoints.net/musical_symbols
	https://codepoints.net/ancient_greek_musical_notation

	https://codepoints.net/mathematical_alphanumeric_symbols

	https://codepoints.net/mahjong_tiles
	https://codepoints.net/domino_tiles
	https://codepoints.net/playing_cards
	https://codepoints.net/enclosed_alphanumeric_supplement
	https://codepoints.net/enclosed_ideographic_supplement

	https://codepoints.net/miscellaneous_symbols_and_pictographs
	https://codepoints.net/emoticons
	https://codepoints.net/ornamental_dingbats
	https://codepoints.net/transport_and_map_symbols
	https://codepoints.net/alchemical_symbols
	https://codepoints.net/geometric_shapes_extended
	https://codepoints.net/supplemental_arrows-c
	https://codepoints.net/supplemental_symbols_and_pictographs

Note this is not intended as a fully comprehensive list, just a list of
common Emojis that were added to the Unicode spec in recent years.