print # files to edit message to stderr
Currently there's no way to avoid the
X files to edit
message when starting vim with more than one file to edit.
Except when vim runs in silent mode, but that has further implications.

The 'shortmess' option is not yet initialized, so it cannot be used
here. Therefore print the message to stderr instead of stdout.

Then stderr can be redirected to avoid showing the message.