Use a session file for the :gui command
Move write_session_file() from gui_gtk_x11.c to misc2.c (with a small
modification which removes the usage of g_strconcat()).

If a user executes `:gui` command, save the session to a temporary file
and execute gvim.exe with the `-S` option to restore the session.
If a user executes `vim.exe -g`, don't use a session file.