win32: Fix Ctrl-X mapping in Visual mode on vim.exe
`:help dos-standard-mappings` describes that Ctrl-X in Visual mode is mapped
to `"*d`.  It is correct on gvim.exe, however, not correct on vim.exe.  It is
mapped to `"-d` on vim.exe.  So, Ctrl-X doesn't cut the selected text into the
clipboard.

I think this is wrong and Ctrl-X should be mapped to `"*d` even on vim.exe.