mac: Fix Command-V mapping in Insert mode
As `:help mac-standard-mappings` says, currently Command-V in Insert mode is
mapped to `<C-R>*`, however this should be mapped to `<C-R><C-O>*`.

On Windows, Shift-Insert was mapped to `<C-R>*` at first.  But it was changed
to `<C-R><C-O>*` in Vim 6.0.  I think that the mapping on mac should be also
changed.

Actually, I'm not a mac user.  I want to hear mac users' opinion.