Delete swap file after asking for permission
When vim crashes and is able to recover edits from a swap file, the swap file is left on the file system. Vim advises to remove it, but this leaves a burden on the user to find and delete the file.

This code change asks the user if the swap file should be deleted. If the user selects that, the swap file is deleted.