Revise resource handling
Move the common part of the resources from the exes to the dll.
* icon files (except the main icon)
* bmp files
* print status dialog

Adjust the handling of HINSTANCE for that.  Move SaveInst() from
gui_w32.c to os_mswin.c and change s_hinst to g_hinst.
Call SaveInst() from DllMain().

Update the dependencies of the resource files in the makefiles.

Embed the manifest to the resource also for vim.exe so that the print
dialog will be shown in XP style.

Notes:
It might be better to separate the resource file for gvim.exe and
vim.exe, because currently "OriginalFilename" in the version resource is
set to "VIM.EXE".

The print dialogs are different between gvim.exe and vim.exe.  Is it
intentional?