suppress # files to edit message when --not-a-term is specified
Currently there's no way to avoid the
X files to edit
message when starting vim with more than one file to edit.
Except when vim runs in silent mode, but that has further implications.
(i.e. doesn't work with vimdiff)

The 'shortmess' option is not yet initialized, so it cannot be used
here. Therefore allow to suppress the message with --not-a-term.