Handle negative values in urxvt mouse termcodes
Urxvt can report negative values for rows and cols in a mouse event termcode (for example, in the case of a mouse drag that extends past the top or left side of a terminal window). The negative sign in such a termcode was preventing it from being properly matched (like ^[64;0;-1M, meant to be matched by ^[64*M); This skips over the negative sign and matches properly.