msvc: Delete Make_dvc.mak and Make_ivc.mak
Both files are now obsoleted and cannot be used in the recent Visual
Studio.

Make_dvc.mak was for debugging, but recent Visual Studio can open
an exe file as a project, and can debug directly.  No need to use it
anymore.

Make_ivc.mak was for VC 4.0 - 6.0 IDE.  Recent Visual Studio cannot
open it anymore.