win32: vimdll: Some error messages are not shown on gvim.exe
Some error messages output via `mainerr()` in main.c are not shown or just
shown as informative messages when gvim is compiled with VIMDLL.

When `mainerr()` is called from `early_arg_scan()`, messages are not shown
at all.  This is because both `gui.starting` and `gui.in_use` are FALSE,
so `mch_errmsg()` tries to output the message to stderr and a dialog box
is not shown.
E.g.: `gvim --servername` shows nothing.
It is the same for the `ole_error()` in gui_w32.c.

When `mainerr()` is called from `command_line_scan()`, messages are shown
as non-error messages.  This is because `gui.starting` is set, and
`display_errors()` in os_mswin.c checks it.  When the flag is set, the
dialog box is shown as an informative dialog, not an error dialog.
E.g.: `gvim --windowid` opens an informative dialog.

To work around them, this PR sets `gui.in_use` and unsets `gui.starting`
at the top of the `mainerr()`.

This also fixes that `gvim --serverlist` opened an error dialog when
compiled without VIMDLL.  It is not an error, so it should be shown
as an informative dialog.  Set `gui.starting` earlier (before
`exec_on_server()`) on non-VIMDLL to show as an informative dialog.

Also fixes wrong conditions in messages.c.
E.g.: `gvim --servername` shows nothing because of this even without
VIMDLL.