Use the last message in python exception as the Vim exception
Python exception traces are printed bottom-up, with the actual exception
type and message last. However, the way that 'emsg' works is that when
it decides to raise an exception, it treats the _first_ message in a
series of messages within a try as the exception text.

So for python, when printing out the exception trace (or any other error
for that matter), we tell emsg that the message is severe. This forces
it to keep replacing the "exception" message with the traced message.
The last such message (i.e. the one we want) ends up in the Vim exception
message.

Concretely: previously, v:exception would always be "Traceback, most
recent call last", now it's something more like "RuntimeError: this is
the error message", which is much more useful.

This is particularly useful when running Vim's own tests, or tests based
on that, where the actual emesg is suppressed due to the try/catch
handling of the test infrastructure.