Adjust popup position to the left if run out of space
When the popup is left-aligned and run out of screen columns to display
it, and wrapping is not enabled, adjust the position left in order to
disaply as much of the popup contents as possible, much like the
completion PUM.

this is particularly useful for argument hints becuase it allows the
maximum amound of hinting visible while remaining
one-signature-per-line.

Maybe this is controversial, so I guess discuss.