Add popup_settext
The {text} argument to popup_create is a useful way to specify the
contents of the buffer. In order to minimise buffer churn, we allow the
same mechanism (and automatic sizing) to be used when changing the
buffer contents. this allows for simple code which creates the buffer
contents/properties as a dict and then calls popup_create or
popup_settext depending on whether or not a popup winid already exists.