Fix sign icons not displayed correct when foldcolumns is not 0
v8.1.1587 / #4578 introduced right-aligning of sign icons to the number
columns, but it didn't properly take the original "col" variable into
account and initialize it to 0. This disregards the fact that "col" can
be non-zero, e.g. when foldcolumns != 0.

To see this issue, simply call `set foldcolumn=2` and then create a GUI
sign icon. Note that the icon is not correctly pushed to the right by 2
columns and therefore overlap the fold column region. This fix will make
sure that's taken into account.