Fix flaky test_suspend script test
Sometimes, test_suspend would seem to fail in CI for GUI tests in
MacVim, when it's waiting for the shell to quit in
`StopShellInTerminal()`. The issue seems to be that the test just calls
`term_wait(buf)` to make sure the key has been handled, but it doesn't
actually properly wait for the shell to come back up (and `term_wait`
has a pretty short default wait time) before issuing the "exit" command,
leading to some odd race conditions.

Add a wait to make sure the shell prompt is up, before issuing the
"exit" command. This helps remove potential race conditions.