Fix sign icons not aligned properly with cursorline
A previous change (#4627) fixed an issue where sign icons were not
displayed correctly with non-zero foldcolumn setting, but it introduced
a small regression where it sometimes causes the sign icon to be
mis-aligned by one column if cursorline is on, and the cursor is on the
line where there's a sign icon.

Fix the regression by making sure we assign `signcol` to `col` only
after it's been fully calculated.