:args: do not add trailing newline(s)
Since v8.0.1738 (5d69da462) `:args` with only a single argument would
cause a hit-enter message:

    [/etc/fstab]

    Press ENTER or type command to continue

This patch changes it to not add a trailing newline.

It adds `crow` for "current row", which could be calculated as
`last_row` instead similar to `last_col` maybe, but I found this easier,
and the patch is only meant as a suggestion and will be edited anyway.

Likely also needs a test adjustment then (or should have a test).