Force redraw for appendbufline()
This will force a redraw for set_buffer_lines, if it has been modifying
a non-current buffer by temporarily disabling RedrawingDisabled and
forcing an redraw.

The reason for this is, that a user might expect a redrawn window when
using appendbufline/setbufline for a different buffer. Currently this
only happens when the cursor switches to it.

Add a screen dump test to verify its behaviour.

fixes #4718