Correctly use count for g$ when test is not wrapped
This fixes issue #4566 the reason being, that the count was not taken
correctly into account when the text file was not being wrapped.

So when the test is not wrapped, make the cursor go down count-1 lines
and then go to the end of the visual line.

Add a couple of tests to the current test_normal.vim to verify this.