Add keyboard layout info for typing $
Turkish language is commonly typed by two different keyboard layouts, Q and F. Therefore some symbols are typed by different keyboard combinations. In this case $ is typed by ALT GR + 4 in Turkish Q, and SHIFT + 4 in Turkish F. Original text mistakenly indicated that ALT + 4 would be sufficient, but that is not the case.