Correctly reset highlighting after wrapping
If the character continues on the next line, need to correctly reset the
character highlighting so that screenline cursorline is not applied to
the wrapped part of the character. So save and store the previous
char_attr before applying the cursorline highlighting to it.

Once the line wrapps, we need to test if wrapping occurs and a screen
line cursorline highlighting is needed and apply the stored highlighting
rather than the actual char_attr to the next chars.

Also: correctly handle diff mode

Make sure to use the line highlighting and combine it with the
cursorline highlighting attribute.

In addition, correctly highlight parts of previous line
that have wrapped