Apply screenline cursorline to special chars
First of all, we need to add the cursorline screen highlighting to
special items on a line (e.g. like special multibyte characters like
 `<200d>` zero-width joiner or to a tab when list mode is set `^I`.

However this might pose a problem, if this character spans over a line
and is wrapped. Then the cursorline highlighting will be extended on the
next screen line. To prevent this, save and restore the extra_attr
before applying the cursorline highlighting to it and restore it after a
line has been wrapped.

Also add screenline cursorline highlighting in a couple of more places