Correctly handle precedes char in list mode for long lines
The 'precedes' value for the listchars option should only be drawn when
the list option is on and wrapping is off (and there is text preceeding
the first column) OR when list option is on and wrapping is on and the
line is longer than the complete window (and there is still text
preceeding the first column). But in the latter case, only
draw the precedes value in the first column visible for the physical
line, not in every first column of the screen line.

Add a test to verify the behaviour.

related: neovim/neovim#11034

(Note: the change to view_util is actually unrelated, but makes debugging
easier as the window size does not change unexpectedly when trying to
run the test interactively and switching between windows).