Extend filetype detection to shebangs which use env -S
As of GNU coreutils 8.30 (July 2018), env(1) supports
an -S/--split-string option. This allows multiple parameters
to be passed to a command rather than processing the whole
command as a single argument. This is also supported on
FreeBSD, which inspired the feature. [1]

This patch allows Vim to detect (i.e. skip/ignore) this option
so that filetype detection continues to work for shebangs
which use this option on Linux/FreeBSD.

[1] https://lists.gnu.org/r/coreutils/2018-04/msg00011.html