usr_03.txt: Attempt to avoid a misunderstanding
I can imagine the previous wording being confusing since it has repeated
the word "string". It could look like "/string" itself is the command to
search for a string, and maybe "/number" would be to search for a
number, and so on. I think introducing it as "/{pattern}" or perhaps
"/pattern" would be better.