Correct test.
When visually selecting text using `v` and using `gn` the selection
should not wrap around.

That means in a text like this:

abcdefhji

selecting the `h`, and the 2 chars in front of it using e.g. `fhvhh`
and then using `gn` with a search pattern of `b` should NOT select
anything between `b` and `h`.

That is, because `gn` starts a search in forward direction and does not
wrap around. It worked previously, because resetting the `wrapscan`
property was done to early and so the `gn` did previously wrap around
and select the b. However as this depends on the initial value of the
wrapscan property and `gn` is documented to search forwards (down the
buffer), I don't think it makes sense to expect it to start selecting
anything from the start of the buffer.

Therefore change the test case to the expected result