gn: Rename is_one_char function to zero-width
The initial intent of this function was to only catch zero-width match
width pattern. However, when selecting zero-width matches it behaves
visually the same as when selecting matches that are exactly one char
wide.

However since patch v8.1.0629 (commit 5d24a2257e597fd7) the
current_search() function does now know the end position of the match,
so it is not necessary anymore to handle one-character width matches
specifically.

Therefore rename the function to make sure the intent is to detect
zero-width matches.