gn: correctly handle start-of-selection in visual_mode
When selecting backwards and the cursor is at the beginning of the
visual selection and VIsual marks the end, put the cursor on the start
of the match, instead of the end-match after the search has been
performed.

e.g. with a text like this one (and `5` is in the search register)

12345678

and selecting using `v` using `:norm! 0f7vhhhh` (such that everyting
from the 3 till the 7 is selected and the cursor is at the start of the
selection), pressing `gn` should make the selection from `5` till `7`
inclusive, that is the cursor will be moved to the start of the match.