:reg: display register type for :reg command
This has been wished for in  vim/vim#4546

Register type can be one of linewise, charwise or blockwise. So when
outputting the register contents, also indicate in the first column the
type of the register using 'l' for linewise, 'v' for charwise and 'b'
for blockwise.

To be consistent with getregtype(), it might have been better to use
`v`, `V` and `^V` as indicators for the type, however I personally find
those types rather hard to distinguish in the console output, so I took
the liberty to use letters that are more easily distinguishable.

the test test_register.vim has been adjusted  and the documentation
amended.