Fix some small MATLAB indenting issues
Fix MATLAB indent at start of line:

For the first non-blank line of the file `prevlnum` would be `-1`
meaning that `indent(prevlnum)` would also be `-1`, which would get
returned by `GetMatlabIndent()`. Thus indentation was left untouched.

Fix indenting of `end` in cell array indexing:

The previous `end` heuristic only accounted for indexing using
parentheses, for example:

    A(2:end) = 0;

However if `end` was used when indexing a cell array, it would be
falsely be picked up as an end statement, causing following code to be
dedented:

    if true
        cellarray{2:end} = A;
    disp foo
    end

This fixes the issue by adding `}` to the regular expression used as the
heuristic.

Fix MATLAB indent not treating `%%` as comment:

The section title after `%%` was considered a different syntax item than
regular comments by the Vim MATLAB syntax file, and thus not ignored
when calculating indentation.