syntax: make: don't generally parse strings in make code
Apparently make does not have any notion of strings, and everything that
is after a = or := will be assigned literally to the variable before it,
even single and double quotes [1]. The only exceptions I could find are
quoting included files in BSD make [3] (which we took care of in the
previous change), and quoting of long file names with double quotes in
Microsoft NMAKE targets [2]. Also the commands are executed in a shell,
which usually has a notion of strings with single and double quotes.
Restrict parsing of strings to those two ocurrences.

[1]: https://stackoverflow.com/a/23332194
[2]: https://docs.microsoft.com/en-us/cpp/build/reference/long-filenames-in-a-makefile?view=vs-2019
[3]: https://www.freebsd.org/cgi/man.cgi?make%281%29, Section "INCLUDESTATEMENTS, CONDITIONALS AND FOR LOOPS"