Add extended version of readdir()
This adds a function named readdirex() which is an extended version of
the readdir() function.
This retrieves a list of files in a directory along with their attributes.

This might be useful when implementing file managers like Netrw.

This is much faster than calling readdir() to get a list, then calling
getfperm(), getfsize(), getftime() and getftype() for each file to get the
attributes, especially on MS-Windows.