Graduate FEAT_NUM64
Since 8.0.1736, support for long long has been required, so we should
always have support for 64-bit values.

Additionally, the conditional support for non-64 bit vimscript variables
causes problems with other parts of the code that are 64-bit even when
FEAT_NUM64 is disabled.

For example, on mipsel a long is 32-bit but off_t is 64-bit.  Internal
to vim_vsnprintf_typval(), Vim is converting %lld to %ld even when the
type truly is 64-bit.  Depending on the alignment (and value) of the
variable being formatted, the value may be stringified incorrectly.

Always building with FEAT_NUM64 avoids these problems