Make running individual tests more robust for environment overrides
Right now, to run an individual test, you can run `make test_cmdline` in
testdir which will clean the .res file and then basically call `make
test_cmdline.res`. However, it does not pass through misc variables like
"VIMPROG" that could be overriden in command line or by the parent
Makefile in src. Pass that through so that the behavior is consistent.

Also, include instructions in the README inside testdir for how to run
individual tests in GUI.