Swift: Change tabstop and shiftwidth to 4
Due to Swift project's code indentation settings, Swift filetype plugin uses 2 spaces for `ts` and `sw`. This conflicts with Xcode, which uses 4 by default (and generally Swift projects use 4), and if same file is opened via Vim or vice versa, it's always extra steps to fix indentation. This PR sets `ts` and `sw` to 4, for a more unified environment.

I've brought this up to Swift team, but they prefer to keep it 2.