Fix error with overlapping props with same id/type
There is nothing in the documentation that forbids having multiple text
properties that all share the same id and type combination, therefore it
seems strange that it would cause errors upon deleting/joining lines.

Disambiguating which prop belongs to which on the other line is
impossible without ensuring property id uniqueness, but the result will
at least always be visually correct when joining, i.e. the whole union
of the text covered by the props will be accounted for by the joined
props.

Could also consider looping backward over the props with

    for (done_this = this_props_len; (done_this -= sizeof(textprop_T)) >= 0; )

since they are sorted by starting column.