Added a tag property to the job_info object.
With this it is possible to supply additional parameters to the exit
callback in exit_cb via a dictionary. A simple use case would be to echo
a message in the exit callback that tells the user which job has just
exited.

Using the job as the key does not work for very short lived jobs, because
the exit handler could already be called before the initiating caller has
the chance to place the parameters in the dictionary.