Fix Buffers menu not adding new buffers to menu
v8.2.0413 attempted to fix an issue with special buffers leaving stale
Buffers menu items. Unfortunately, it had a bug where it use passing a
string to `buflisted()` and `bufexists()` for checking whether buffer
can be added to menu which leads to the logic not working (as those
functions behave differently on string vs number types). As a result
*no* buffer is currently added to the Buffers menu. Fix this by doing a
number type coersion by doing a "+ 0".

A caveat of this is that the logic to check special buffers (using
`buftype`) actually does not work, because the code triggers the BufAdd
autocommand event before setting the buftype, meaning it's always ''
when we handle the autocmd.

Also, update tests so that:
1. Check that making a new buffer actually adds a new item to Buffers
   menu to prevent future regression.
2. Fix up Test_menu_expand() to behave properly again by calling
   delmenu.vim at beginning of test. We have to do this because even
   though we clearn up by calling delmenu.vim in earlier tests, that's
   not sufficient, as delmenu.vim doesn't actually clear the Buffers
   menu's autocmd (BMAdd).
3. Fix up the test to expect that command-line window and terminal
   windows *do* in fact create new buffer menu items, because of the
   'buftype' bug that I mentioned above.