doc: Mention that getcurpos always uses bufnr=0
For #1621.

Reading the docs for getcurpos, you'd expect that the first element in
the list is the current buffer number and you'd have to check the help
for getpos to see what's gone wrong.

I'm not really clear on *why* we use 0, so I think this is nonobvious
enough that it merits a mention here too.