term_util: Add TermWait() to apply the pattern from 8.2.0454 everywhere
Since interacting with jobs spawned in a terminal is generally more
sensitive to what's happening on the system, provide a wrapper around
term_wait() to automatically wait longer on retried test runs.

The wait time, either term_wait()'s default or the user-specified, will
be multiplied by 4 on the first retry and 10 on any subsequent retries,
as was initially done in 8.2.0454.