Update docs of i_CTRL-R_CTRL-R
The documentation of i_CTRL-R_CTRL-R, i_CTRL-R_CTRL-P and
i_CTRL-R_CTRL-O didn't include an explanation about what stays in the
'.'
register after them, and worked differently.

The command of i_CTRL-R_CTRL-R works the same as the command i_CTRL-R in
respect to this register, but the commands of i_CTRL-R_CTRL-P and
i_CTRL-R_CTRL-O inserts the commands to the '.' register, and not the
text produced as a result of this command.

This commit adds this documentation to the documentation in the help of
these commands.

Fixes #5771.