Fix documentation of v_r command
In the documentation of the v_r command, it says that the command can
get a register. However, the command does nothing with this
register. From the current documentation, it seems that the command
would insert the text into the given register (like the command of v_s
or v_c does), but in practice ,the command does nothing with this
register.

Fixes #3247.