findfile - correct result order
Previously findfile was using stack of directories; This means that the
order of directories which where searched was reveresed makeing it
unintuitive to use.  Instead, we use doubly-linked list, to which we
push elements to the front and pop elements from its tail (what
guarentees FIFO order).

This is potentially breaking change, if a script relies on the order of
found files.

Fixes issue #5888