Fix wrong color assignation
Most terminal emulators allow users to customize their colors by
selecting a predefined palette and/or by picking individual colors.

This means that colors 0-15 can't be expected to have a fixed value and,
therefore, that no attempt should be made to map them to hardcoded ones.

With "&t_Co=256",

* "darkred" is correctly mapped to "1",
* "darkgreen" is correctly mapped to "2",
* "darkblue" is correctly mapped to "4",
* "darkmagenta" is correctly mapped to "5",
* "darkcyan" is correctly mapped to "6",

but "darkyellow" is incorrectly mapped to "130", a brownish orange,
instead of the expected "3".

Cause: "darkyellow" is mapped to xterm color 130

Solution: map "darkyellow" to user-customisable xterm color 3