Add support for separate underline coloring
This patch introduces the new terminal output codes 't_AU' and 't_8u'
and a new highlight argument 'ctermul'.  These allow coloring of
underline and undercurl independently from the foreground color.

If 'termguicolors' is set, 'guisp' will be used to color the underline.

If 'ctermul' or 'guisp' are not specified behaviour remains the same,
which is using the foreground color for underline.