Handle ignored signals correctly
If Vim is started with some signals ignored they should remain ignored.
This is particularly important for the tty job control signals such as
SIGTSTP which might be marked as ignored by a shell that does not
implement tty job control.

I verified this on macOS and Linux with bash (which implements tty job
control) and elvish (which does not).

Lastly, this fixes a bug in test_channel_pipe.py that can leave spinning
processes when `make test` is run. Sorry for including an unrelated change
but it's two lines and keeps systems from accumulating spinning versions
of that program when `make test` is run.

Fixes #5990